few tweaks to URL configs
