+++ straxen/__init__.py
@@ -1,11 +1,10 @@
 __version__ = '1.8.3'
 
-from utilix import uconfig
 from .common import *
 # contexts.py below
 from .corrections_services import *
 from .get_corrections import *
-from .hitfinder_thresholds import *
+from .legacy.hitfinder_thresholds import *
 from .itp_map import *
 from .matplotlib_utils import *
 from .mini_analysis import *
@@ -20,7 +19,7 @@ from .plugins import *
 
 from . import storage
 from .storage import *
-from .legacy_fdc import *
+from .legacy.legacy_url_configs import *
 from . import analyses
 
 # Do not make all contexts directly available under straxen.
