+++ straxen/legacy/README.md
@@ -1,0 +1,2 @@
+# 1T support files
+In this subdirectory, we store XENON1T related files / scripts that are only used for xenon1t
